Have WebProcessPool::processForRegistrableDomain() return a Ref<WebProcessProxy> https://bugs.webkit.org/show_bug.cgi?id=222575
Reviewed by Geoffrey Garen.

Have WebProcessPool::processForRegistrableDomain() return a Ref<WebProcessProxy> instead of a WebProcessProxy&,
for extra safety. Returning a WebProcessProxy& is generally safe because the WebProcessPool is expected to hold
a reference to the WebProcessProxy it is returning (via WebProcessPool::m_processes). However, this is a bit
fragile and can lead to bad bugs as we've seen recently.

* UIProcess/WebProcessPool.cpp:
(WebKit::WebProcessPool::createNewWebProcess):
(WebKit::WebProcessPool::processForRegistrableDomain):
(WebKit::WebProcessPool::createWebPage):
* UIProcess/WebProcessPool.h:


Canonical link: https://commits.webkit.org/234714@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@273693 268f45cc-cd09-0410-ab3c-d52691b4dbfc