Adjust progress parameter before calling blend() for discrete interpolations https://bugs.webkit.org/show_bug.cgi?id=222736
Reviewed by Antti Koivisto.

In the case of discrete interpolation, we must treat progress as 0 for progress < 0.5
and 1 otherwise. Currently we do it separately in at two places. So that we don't do
it in more places, let's do this directly in CSSPropertyAnimation::blendProperties()
so that the progress is already set before any of the blend() methods are called.

* animation/CSSPropertyAnimation.cpp:
(WebCore::CSSPropertyAnimation::blendProperties):

Canonical link: https://commits.webkit.org/234857@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@273896 268f45cc-cd09-0410-ab3c-d52691b4dbfc