Add internal preference to disable HTTPS upgrade https://bugs.webkit.org/show_bug.cgi?id=222778
Patch by Alex Christensen <achristensen@webkit.org> on 2021-03-04
Reviewed by Simon Fraser.

Source/WebKit:

* UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::creationParameters):

Source/WTF:

This is needed for an internal performance benchmark, which serves canned content from a local http server
pretending to be from different domains that are in our internal HTTPS upgrade list because the real server
supports HTTPS.  That internal benchmark should eventually change, but that's not going to happen this year.
This is also likely going to be useful for QA purposes to be able to easily tell on an internal build
whether HTTPS upgrade is breaking something.

* Scripts/Preferences/WebPreferencesInternal.yaml:

Canonical link: https://commits.webkit.org/234898@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@273950 268f45cc-cd09-0410-ab3c-d52691b4dbfc