Use WTF::DestructionThread::MainRunLoop for CallbackAggregators in WebsiteDataStore https://bugs.webkit.org/show_bug.cgi?id=222831
Reviewed by Geoffrey Garen.

Use WTF::DestructionThread::MainRunLoop for CallbackAggregators in WebsiteDataStore to make sure
the CallbackAggregators get destroyed on the main thread. Previously, the code was letting the
CallbackAggregators get destroyed on any thread and then dispatching to the main thread in their
destructor their WebsiteDataStore data member to make sure it gets destroyed on the main thread.

It is simpler and safer to have the CallbackAggregators get destroyed on the main thread.

Also add threading assertions whenever the pendingCallbacks is used since this counter is not
atomic.

* UIProcess/WebsiteData/WebsiteDataStore.cpp:
(WebKit::WebsiteDataStore::removeData):


Canonical link: https://commits.webkit.org/234954@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@274018 268f45cc-cd09-0410-ab3c-d52691b4dbfc