dlopen_preflight is failing (temporarily) but obsolete https://bugs.webkit.org/show_bug.cgi?id=222829
Reviewed by Jer Noble.

I hit a bug where dlopen_preflight was failing in a particular build.
That is now fixed, but as I was discussing it with the dyld team
they said we should stop using it. The rationale was that it only
made sense during the PowerPC to Intel transition, and it is as
expensive as dlopen now.

Source/WebCore:

* platform/graphics/avfoundation/objc/AVAssetMIMETypeCache.mm:
(WebCore::AVAssetMIMETypeCache::isAvailable const): Remove the preflight opening
of the AVFoundation framework.
* platform/graphics/avfoundation/objc/AVStreamDataParserMIMETypeCache.mm:
(WebCore::AVStreamDataParserMIMETypeCache::isAvailable const):

Source/WTF:

* wtf/PlatformUse.h: No need for preflight linking code.
* wtf/cocoa/SoftLinking.h:

Canonical link: https://commits.webkit.org/234970@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@274034 268f45cc-cd09-0410-ab3c-d52691b4dbfc