Crash under WebViewImpl::flagsChanged() https://bugs.webkit.org/show_bug.cgi?id=222989
Reviewed by Tim Horton.

Capture a WeakPtr for |this| in the lambda passed to interpretKeyEvent() and make sure |this|
is still alive before getting m_page. This is the same pattern that is used for other
interpretKeyEvent() calls in this file (but this one was missed).

* UIProcess/Cocoa/WebViewImpl.mm:
(WebKit::WebViewImpl::flagsChanged):


Canonical link: https://commits.webkit.org/235097@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@274176 268f45cc-cd09-0410-ab3c-d52691b4dbfc