Nullptr crash in Node::isTextNode() via ApplyBlockElementCommand::endOfNextParagraphSplittingTextNodesIfNeeded https://bugs.webkit.org/show_bug.cgi?id=222620
Patch by Venky Dass <yaranamavenkataramana@apple.com> on 2021-03-09
Reviewed by Ryosuke Niwa.
Source/WebCore:

When a text node does not have a previous sibling there is a crash. The fix is to check if there is a previous sibling
before we try to de-reference it.

Test: editing/inserting/indent-split-text-not-having-previous-sibling-crash.html

* editing/ApplyBlockElementCommand.cpp:
(WebCore::ApplyBlockElementCommand::endOfNextParagraphSplittingTextNodesIfNeeded):

LayoutTests:

Adding a regression test.

* editing/inserting/indent-split-text-not-having-previous-sibling-crash-expected.txt: Added.
* editing/inserting/indent-split-text-not-having-previous-sibling-crash.html: Added.

Canonical link: https://commits.webkit.org/235118@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@274200 268f45cc-cd09-0410-ab3c-d52691b4dbfc