Release assertion failures under Editor::scanSelectionForTelephoneNumbers https://bugs.webkit.org/show_bug.cgi?id=223016 <rdar://problem/73159921>
Reviewed by Ryosuke Niwa.

No new tests; speculative fix for a non-reproducible crash, which theoretically
has been avoided in a second way on trunk.

* editing/Editor.cpp:
(WebCore::extendSelection):
(WebCore::Editor::scanSelectionForTelephoneNumbers):
In r272777, Ryosuke discovered a case where FrameSelection::isRange()
can be true, but firstRange() is an invalid range; in testing, forcing
this to be the case reproduces the crash as reported.

While that change may have fixed the root cause of this crash, we don't
know that we've found all of the ways that one can get a orphaned
selection *into* FrameSelection, and want a less risky workaround for
this crash, so we'll also fix it in scanSelectionForTelephoneNumbers,
by null-checking the result of FrameSelection::firstRange().

Also null-check the result of extendSelection(), since I cannot prove for sure
that a valid range cannot become invalid in this method.


Canonical link: https://commits.webkit.org/235120@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@274203 268f45cc-cd09-0410-ab3c-d52691b4dbfc