REGRESSION (r267775): Web Share API Level 2 is incorrectly disabled https://bugs.webkit.org/show_bug.cgi?id=223110
Reviewed by Simon Fraser.

* Scripts/Preferences/WebPreferencesExperimental.yaml:
Through a series of minor mishaps related to r267623 and r267775,
the current state of the default value of WebShareFileAPIEnabled
in modern WebKit does not match what it was changed to in r267762
(is it now instead off everywhere).

Fix this regression by copying the condition from WebShareEnabled.


Canonical link: https://commits.webkit.org/235216@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@274328 268f45cc-cd09-0410-ab3c-d52691b4dbfc