[JSC] wasm atomic wait offset is not index https://bugs.webkit.org/show_bug.cgi?id=223159
Reviewed by Mark Lam.

JSTests:

* wasm.yaml:
* wasm/threads-spec-tests/resources/wait-large.wast: Added.
* wasm/threads-spec-tests/wait-large.wast.js: Added.

Source/JavaScriptCore:

While JS Atomics.wait's argument is "index" in the typed-array, argument of wasm wait and notify is address.
But we are handling it as an index incorrectly.
This patch uses it as an address.

* wasm/WasmOperations.cpp:
(JSC::Wasm::JSC_DEFINE_JIT_OPERATION):


Canonical link: https://commits.webkit.org/235266@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@274399 268f45cc-cd09-0410-ab3c-d52691b4dbfc