[GStreamer][MSE] fix video freeze in NASA TV feed https://bugs.webkit.org/show_bug.cgi?id=214625
Reviewed by Alicia Boya Garcia.

Initial seek should not be marked as a real seek as it is not
getting seek-datas from GStreamer and therefore it can mess up
with pipeline branches becoming empty after getting an enough-data
because the need-data is expecting seek-datas to complete that
seek that is never going to happen.

A couple of debug output lines were fixed for the same price.

* platform/graphics/gstreamer/mse/WebKitMediaSourceGStreamer.cpp:
(webKitMediaSrcPrepareInitialSeek):


Canonical link: https://commits.webkit.org/235334@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@274478 268f45cc-cd09-0410-ab3c-d52691b4dbfc