[iOS] Scroll event is fired before webkitfullscreenchange event when an element enters fullscreen https://bugs.webkit.org/show_bug.cgi?id=223221
Reviewed by Eric Carlson.

Defer the "scroll" event until the "webkitfullscreenchange" event is fired
when an element enters fullscreen. We have to do that because some sites use
"scroll" event handler to remove elements out of visible area, which may
conflict with the entering fullscreen process.

No new tests. We cannot test this patch with a layout test because
`InjectedBundlePageFullScreenClient` in WKTR does not support testing the
fullscreen implementation in the UI process, which is related to the "scroll"
event. We cannot add an API test on iOS for it either because the test binary
is not a real "app" (webkit.org/b/212654).

* UIProcess/ios/fullscreen/WKFullScreenWindowControllerIOS.mm:
(-[WKFullScreenWindowController enterFullScreen]):
(-[WKFullScreenWindowController beganEnterFullScreenWithInitialFrame:finalFrame:]):


Canonical link: https://commits.webkit.org/235343@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@274487 268f45cc-cd09-0410-ab3c-d52691b4dbfc