[css-grid] Incorrect track sizing when using relative sized items in 'auto' column tracks https://bugs.webkit.org/show_bug.cgi?id=191627
Patch by Ziran Sun <zsun@igalia.com> on 2021-03-17
Reviewed by Javier Fernandez.

LayoutTests/imported/w3c:

Updated expectation file for test grid-self-baseline-not-applied-if-sizing-cyclic-dependency-001.html.
All sub-tests in this test file are now all passed.

* web-platform-tests/css/css-grid/alignment/grid-self-baseline-not-applied-if-sizing-cyclic-dependency-001-expected.txt:

Source/WebCore:

Containing block for a grid item is the grid area it's located in. During the GridTrackSizingAlgorithm,
we determine that this item has a relative size (percent in this case) and that the grid area is still
indefinite. Hence, we set nullopt for the grid area width and mark the item for layout. During the item's
layout, we call this RenderBox::containingBlockLogicalWidthForContent to determine the available height.
However, this method checks the containing block's width if the overridingContentLogicalWidth is nullopt,
which is incorrect for grid items. We should never use the grid container for resolving grid item's size.

This change is to return 0_lu for grid item when overridingContainingBlockContentLogicalHeight() or
overridingContainingBlockContentLogicalWidth() returns nullopt rather than use containing block's width
or height.

* rendering/RenderBox.cpp:
(WebCore::RenderBox::containingBlockLogicalWidthForContent const):

LayoutTests:

* TestExpectations:

Canonical link: https://commits.webkit.org/235437@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@274596 268f45cc-cd09-0410-ab3c-d52691b4dbfc