should always use all prerequisites for the `ModernMediaControls.{js,css}` rules in DerivedSources.make https://bugs.webkit.org/show_bug.cgi?id=223421
Reviewed by Yusuke Suzuki.

* DerivedSources.make:
`$?` is a uniqued space-separated list of all prerequisites modified after the target was modified.
`$^` is a uniqued space-separated list of all prerequisites regardless of when the target/prerequisite was modified.
We want to use the latter since we always want to `cat` all the JS/CSS files, not just the recently edited ones.


Canonical link: https://commits.webkit.org/235462@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@274625 268f45cc-cd09-0410-ab3c-d52691b4dbfc