Source/WebKit/UIProcess/Cocoa/WebViewImpl.mm:2681:9: runtime error: load of value nnn, which is not a valid value for type 'BOOL' (aka 'signed char') https://bugs.webkit.org/show_bug.cgi?id=223581
Reviewed by Simon Fraser.

The code was implicitly casting an uint64_t identifier into a BOOL variable. The intention was
to check whether the identifier is 0 or not. To silence the UBSAN error, we now use the bool
type instead of BOOL (since BOOL is defined as a signed char on some platforms).

* UIProcess/Cocoa/WebViewImpl.mm:
(WebKit::WebViewImpl::pluginFocusOrWindowFocusChanged):


Canonical link: https://commits.webkit.org/235590@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@274786 268f45cc-cd09-0410-ab3c-d52691b4dbfc