[YARR] Interpreter incorrectly matches non-BMP characters with multiple . https://bugs.webkit.org/show_bug.cgi?id=223498
Reviewed by Yusuke Suzuki.

JSTests:

New test.

* stress/regexp-dot-match-nonBMP.js: Added.
(shouldMatch):
(shouldntMatch):

Source/JavaScriptCore:

We need to check that we read an actual character before seeing if it is part of a character class.
In the case where we are checking that a character is not in a character class, like .,
the failed to read result from input.readChecked(), -1, is not part of the newline character class.
This will allow regular expressions that require more than the number of characters in a string
to match.

* yarr/YarrInterpreter.cpp:
(JSC::Yarr::Interpreter::checkCharacterClass):

LayoutTests:

Updated test.

* fast/forms/ValidityState-patternMismatch-expected.txt:
* fast/forms/ValidityState-patternMismatch.html:


Canonical link: https://commits.webkit.org/235606@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@274806 268f45cc-cd09-0410-ab3c-d52691b4dbfc