REGRESSION(r274919) [ATK] accessibility/set-selected-text-range-contenteditable.html is timing out https://bugs.webkit.org/show_bug.cgi?id=223710
Reviewed by Adrian Perez de Castro.

r274919 clamped the values passed to the accessibility elements,
making the {-1,0} text range to be collapsed to {0,0} as text ranges
are composed of unsigned ints.

This commit updates the test to remove the ATK-specific case of
passing -1 to point to the final offset of a text range.

* accessibility/set-selected-text-range-contenteditable.html: Remove
ATK-specific quirk.
* platform/glib/accessibility/set-selected-text-range-contenteditable-expected.txt: Removed.


Canonical link: https://commits.webkit.org/235729@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@274984 268f45cc-cd09-0410-ab3c-d52691b4dbfc