Improve interpolation of the shape-outside CSS property https://bugs.webkit.org/show_bug.cgi?id=223738
Reviewed by Antti Koivisto.

LayoutTests/imported/w3c:

Add another 29 PASS results for shape-outside interpolation.

* web-platform-tests/css/css-shapes/animation/shape-outside-composition-expected.txt:
* web-platform-tests/css/css-shapes/animation/shape-outside-interpolation-expected.txt:
* web-platform-tests/web-animations/animation-model/animation-types/interpolation-per-property-002-expected.txt:

Source/WebCore:

While we had logic to deal with cases where ShapeValue objects could not be interpolated, we did not make
use of it in the animation wrapper. We now move this logic from the blendFunc() to the canInterpolate()
override method on the wrapper to correctly interpolate only when possible.

* animation/CSSPropertyAnimation.cpp:
(WebCore::blendFunc):

Canonical link: https://commits.webkit.org/235748@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@275015 268f45cc-cd09-0410-ab3c-d52691b4dbfc