Merge r274428 - postprocess-asm/resolve-asm-file-conflicts.rb build failure after upgrading to F34 https://bugs.webkit.org/show_bug.cgi?id=223136
Reviewed by Michael Catanzaro.

When parsing .file assembler directives (for the purpose of
deduplicating the file slots), also accept

  .file "path/to/CWD" "path/to/include"

that seems to be emitted by GCC in some configurations. This also
uses Pathname.cleanpath on the resulting path to canonicalize the
paths. We could use .realpath, but since we only run this on the
paths in a single compilation unit and the first component is
supposed to be the CWD, this both seems unnecessary and would
complicate our selftests.

* Scripts/resolve-asm-file-conflicts.rb: