Allow some image overlay content to render in fully transparent image elements https://bugs.webkit.org/show_bug.cgi?id=223781 <rdar://problem/75886351>
Reviewed by Tim Horton.

Source/WebCore:

Introduce `ImageOverlayController`, and use it to render selections in image overlay content when the image
overlay's host element is fully transparent.

Test: fast/images/image-extraction/mac/select-word-in-transparent-image-overlay.html

* Sources.txt:
* WebCore.xcodeproj/project.pbxproj:
* dom/Document.cpp:
(WebCore::Document::willBeRemovedFromFrame):

Add plumbing to allow `ImageOverlayController` to uninstall its `PageOverlay` when the document is about to be
detached. See `ImageOverlayController::` below.

* editing/SelectionRectGatherer.cpp:
(WebCore::SelectionRectGatherer::Notifier::~Notifier):

Have `SelectionRectGatherer` notify the image overlay controller as well, when selection rects change.

* page/ChromeClient.h:
(WebCore::ChromeClient::needsImageOverlayControllerForSelectionPainting const):

Add a new client hook that returns whether or not we should use `ImageOverlayController` to paint selections.
If not (i.e. we're on iOS, where we use UIKit to draw selections in the UI process), then we'll never install
page overlays, since `selectionRectsDidChange` is effectively a no-op.

* page/ImageOverlayController.cpp: Added.
(WebCore::ImageOverlayController::ImageOverlayController):
(WebCore::ImageOverlayController::selectionRectsDidChange):

When selection rects change, if the selection is inside an image overlay whose host is completely transparent
(or very close to being completely transparent), then remember the selection rects along with the renderer's
background color, and use this to render selection highlights separately, in a `PageOverlay`.

(WebCore::ImageOverlayController::documentDetached):

If the `Document` containing the current rendered overlay selection is detached, immediately uninstall the
current `PageOverlay`. This ensures that the overlays don't persist through navigation.

(WebCore::ImageOverlayController::installPageOverlayIfNeeded):
(WebCore::ImageOverlayController::uninstallPageOverlayIfNeeded):

Helper methods to add or remove the `PageOverlay` from the current `Page`.

(WebCore::ImageOverlayController::willMoveToPage):

Uninstall the current page overlay if needed.

(WebCore::ImageOverlayController::drawRect):

Use the information stored in `ImageOverlayController::selectionRectsDidChange` to render custom selection
rects.

* page/ImageOverlayController.h: Added.
* page/Page.cpp:
* page/Page.h:

Add an `ImageOverlayController` to the `Page`.

(WebCore::Page::imageOverlayController):

Source/WebKit:

Implement a new chrome client hook. See WebCore ChangeLog for more information.

* WebProcess/WebCoreSupport/WebChromeClient.h:

LayoutTests:

Adjust an existing layout test so that it actually checks what it was intended to check; add a new layout test
to cover the changes in this bug.

* fast/images/image-extraction/mac/select-word-in-draggable-image-overlay.html:

Drive-by fix: remove some extraneous imported scripts from this layout test, and additionally simplify the test
so that it doesn't rely on event sender to select text.

* fast/images/image-extraction/mac/select-word-in-transparent-image-overlay-expected-mismatch.html: Added.
* fast/images/image-extraction/mac/select-word-in-transparent-image-overlay.html: Added.

Add a new layout test with a fully transparent image element that covers a div of the same size, with a
background image that is identical to the image element's image. This emulates the behavior of certain websites
that prompted this fallback image overlay rendering codepath in the first place.


Canonical link: https://commits.webkit.org/235820@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@275112 268f45cc-cd09-0410-ab3c-d52691b4dbfc